Make the Render thread wait for all tasks to be done
Previously we were checking the number of unproccessed tasks in the
queue. This is dangerous since the program might finish while the last
few tasks are still running.
Use the fact that our ThreadPool returns a future for each task to make
the Render thread actually wait for task completion.